+++ test/exports.test.js
@@ -3,9 +3,10 @@ import {
 	getConfig,
 	getCommentText,
 	compileWordPatterns,
-	generateQuotes,
-	generatePreview,
-	generateThreadTitle
+	// generateQuotes,
+	// generatePreview,
+	// generateThreadTitle,
+	// setInReplyToQuotes
 } from '../index'
 
 describe('exports', () => {
@@ -14,9 +15,10 @@ describe('exports', () => {
 		getConfig.should.be.a('function')
 		getCommentText.should.be.a('function')
 		compileWordPatterns.should.be.a('function')
-		generateQuotes.should.be.a('function')
-		generatePreview.should.be.a('function')
-		generateThreadTitle.should.be.a('function')
+		// generateQuotes.should.be.a('function')
+		// generatePreview.should.be.a('function')
+		// generateThreadTitle.should.be.a('function')
+		// setInReplyToQuotes.should.be.a('function')
 	})
 
 	it('should export CommonJS', () => {
@@ -26,8 +28,9 @@ describe('exports', () => {
 		Library.getConfig.should.be.a('function')
 		Library.getCommentText.should.be.a('function')
 		Library.compileWordPatterns.should.be.a('function')
-		Library.generateQuotes.should.be.a('function')
-		Library.generatePreview.should.be.a('function')
-		Library.generateThreadTitle.should.be.a('function')
+		// Library.generateQuotes.should.be.a('function')
+		// Library.generatePreview.should.be.a('function')
+		// Library.generateThreadTitle.should.be.a('function')
+		// Library.setInReplyToQuotes.should.be.a('function')
 	})
 })